# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": " Mojo Api",
        "description": "MOJO API Descrição",
        "contact": {
            "email": "wesleymotabr@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://back.mojosolutions.co/api/",
            "description": "Server Mojo"
        }
    ],
    "paths": {
        "/admin/dashboard": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Para gerar os dados necessários para a geração dos gráficos",
                "description": "{O type pode ser: participation, evaluation, initiativeCreated, initiativeCreatedByType initiativeCreatedByTags, interests, skills",
                "operationId": "dashboard",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type"
                                ],
                                "properties": {
                                    "data": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "type": {
                                                    "type": "integer"
                                                },
                                                "filters": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "tag": {
                                                                "type": "text"
                                                            },
                                                            "start_time": {
                                                                "type": "datetime"
                                                            },
                                                            "end_time": {
                                                                "type": "datetime"
                                                            },
                                                            "type": {
                                                                "type": "integer"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "type": "skills",
                                    "filters": {
                                        "tag": "4",
                                        "start_time": "2020-03-13 21:55:38",
                                        "end_time": "2020-05-11 21:55:38",
                                        "type": 1
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "quando as demais opções",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "indicator": {
                                                        "type": "number"
                                                    },
                                                    "period": {
                                                        "type": "datetime"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "data": {
                                            "indicator": 0,
                                            "period": "2020-03-13 21:55:38"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
...
        "/notifications/": {
            "get": {
                "tags": [
                    "notifications"
                ],
                "summary": "Notifications",
                "description": "Retornando as notificações do usuério e removendo os convites desatualizados, ordenados pela ordem de criação",
                "operationId": "notifications",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "notifications": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "type": {
                                                        "type": "integer"
                                                    },
                                                    "user_id": {
                                                        "type": "integer"
                                                    },
                                                    "read": {
                                                        "type": "boolean"
                                                    },
                                                    "event_id": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "type": {
                                                                    "type": "integer"
                                                                },
                                                                "title": {
                                                                    "type": "string"
                                                                },
                                                                "code": {
                                                                    "type": "string"
                                                                },
                                                                "category": {
                                                                    "type": "string"
                                                                },
                                                                "start_time": {
                                                                    "type": "datetime"
                                                                },
                                                                "end_time": {
                                                                    "type": "datetime"
                                                                },
                                                                "location": {
                                                                    "type": "string"
                                                                },
                                                                "description": {
                                                                    "type": "string"
                                                                },
                                                                "img": {
                                                                    "type": "string"
                                                                },
                                                                "value": {
                                                                    "type": "double"
                                                                },
                                                                "user_id": {
                                                                    "type": "integer"
                                                                },
                                                                "tags": {
                                                                    "type": "string"
                                                                },
                                                                "recurrence": {
                                                                    "type": "string"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "notifications": [
                                            {
                                                "id": 2,
                                                "user_id": 1,
                                                "event_id": {
                                                    "id": 7,
                                                    "type": 1,
                                                    "category": "teste",
                                                    "title": " LUNCH ‘N’ LEARN – MERO 2",
                                                    "code": "#1",
                                                    "start_time": "2020-03-03 17:58:25",
                                                    "end_time": "2020-03-04 19:58:25",
                                                    "location": "SALA 2",
                                                    "description": "",
                                                    "recurrence": "none",
                                                    "img": "1",
                                                    "value": 10,
                                                    "user_id": 1,
                                                    "event_id": null,
                                                    "boost": 0,
                                                    "created_at": null,
                                                    "updated_at": null
                                                },
                                                "type": 2,
                                                "read": 0,
                                                "created_at": "2020-03-06 21:00:36",
                                                "updated_at": "2020-03-06 21:00:36"
                                            },
                                            {
                                                "id": 1,
                                                "user_id": 1,
                                                "event_id": {
                                                    "id": 15,
                                                    "type": 1,
                                                    "category": "1",
                                                    "title": "sasaasa",
                                                    "code": "#a101",
                                                    "start_time": "2019-12-28 20:45:12",
                                                    "end_time": "2019-08-30 23:05:12",
                                                    "location": "Carrin",
                                                    "description": "legal",
                                                    "recurrence": "none",
                                                    "img": null,
                                                    "value": 10,
                                                    "user_id": 1,
                                                    "event_id": null,
                                                    "boost": 0,
                                                    "created_at": "2020-03-06 20:59:04",
                                                    "updated_at": "2020-03-06 20:59:04"
                                                },
                                                "type": 1,
                                                "read": 0,
                                                "created_at": "2020-03-06 20:59:04",
                                                "updated_at": "2020-03-06 20:59:04"
                                            }
                                        ]
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
...
    },
...
}